ci: build and publish ARM64 images (experimental)
This commit modifies the CI code (`ci.yaml`) to use the new docker build
infrastructure (`Makefile`) in `docker/` to build linux/arm64 images in addition
to linux/amd64.

Several considerations:

* The original step 3 is split in:
  - 3.1: where the basic (local) build for the host arch and the docker-compose
         smoke test is run. This is to mimic what a user would do locally.
  - 3.2: where the multi-platform images are built, smoke tested (host arch only
         and published.
* Due to the limitations of `docker buildx` (see previous commit and
  `docker/README.md`), a temporary build docker registry is spawned as part of
  step 3.2.
* docker image publishing is moved to step of 3.2 (`docker-build-test-publish`),
  which allows PRs to execute most of the sections of 3.1/3.2 and skip the
  publishing step only.

ARM64 generation is slow. Building and - especially - pushing multi-platform
images is much easier than _trying_ to do that across jobs. There isn't much
to optimize on the ARM side of things (just a byproduct of cross-compiling).
However, stage 3.2 could start in parallel with stage 1/2, but gate publishing
to dockerhub until stage 1,2 and 3.1 have successfully completed (e.g. using
artifacts to pass status). This is out of scope of this commit.

Signed-off-by: Marc Sune <marcdevel@gmail.com>